before last commit

+++ app/src/main/java/project/android/todoapp/ui/main/main_screen/MainFragment.kt
@@ -1,8 +1,10 @@
 package project.android.todoapp.ui.main.main_screen
 
 import android.content.Context
+import android.content.DialogInterface
 import android.os.Bundle
 import android.view.*
+import androidx.appcompat.app.AlertDialog
 import androidx.coordinatorlayout.widget.CoordinatorLayout
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
@@ -114,8 +116,19 @@ class MainFragment : Fragment() {
             adapter = adapterTask
             val swipeHandler = object : SwipeToDeleteCallback(context) {
                 override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-                    val adapter = this@apply.adapter as TaskAdapter
-                    adapter.deleteItem(viewHolder.adapterPosition)
+                    AlertDialog.Builder(requireContext())
+                        .setTitle("ToDoForDev")
+                        .setMessage("Do you want to delete this task?")
+                        .setPositiveButton("Yes"
+                        ) { _, _ ->
+                            val adapter = this@apply.adapter as TaskAdapter
+                            adapter.deleteItem(viewHolder.adapterPosition)
+                        }
+                        .setNegativeButton("No"){
+                            _,_ ->
+                        }
+                        .show()
+
                 }
             }
             val itemTouchHelper = ItemTouchHelper(swipeHandler)
